requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.3
boto3==1.34.117
botocore==1.34.117
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.14.0
fonttools==4.47.2
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.25.2
ipython==8.16.1
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.0
openai==1.30.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-transformers==1.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
rdflib==7.0.0
regex==2024.5.15
requests==2.31.0
s3transfer==0.10.1
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.3.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SPARQLWrapper==2.0.0
stack-data==0.6.3
sympy==1.12.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.41.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wandb==0.17.0
wcwidth==0.2.8
zipp==3.17.0
